Refactor: Moved TestBoard to TestClasses file
